src/backend/text_processor.py
# Copyright 2024 fukuro-kun
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

from enum import Enum
from typing import List, Dict, Any, Optional, Tuple
import re

class TextTag(Enum):
    TEXT = "text"
    POSSIBLE_NUMBER = "moeglichezahl"
    LARGE_NUMBER = "grossezahl"
    NUMBER = "zahl"
    IP = "ip"
    DATE = "datum"

class TextPart:
    def __init__(self, content: str, tag: TextTag, position: int):
        self.content = content
        self.tag = tag
        self.position = position
        self.value: Optional[int] = None

class TextLevel:
    def __init__(self, level: float):
        self.level = level
        self.parts: List[TextPart] = []

    def add_part(self, part: TextPart):
        self.parts.append(part)

class TextProcessor:
    def __init__(self):
        self.levels: Dict[float, TextLevel] = {}
        for i in range(5):
            self.levels[i] = TextLevel(i)
        self.current_position = 0
        print("TextProcessor initialisiert")

    def process_text(self, text: str) -> str:
        print(f"Verarbeite Text: {text}")
        self.current_position = 0
        self.split_special_formats(text, 0)
        self.print_levels_state()

        self.identify_number_words()
        self.print_levels_state()

        self.process_number_words()
        self.print_levels_state()

        self.accumulate_numbers()
        self.print_levels_state()

        result = self.reconstruct_text()
        print(f"Verarbeitetes Ergebnis: {result}")
        self.reset_levels()
        return result

    def reset_levels(self):
        for level in self.levels.values():
            level.parts = []

    def split_special_formats(self, text: str, level: float) -> None:
        print(f"Teile spezielle Formate auf: {text}")
        self.levels[1].parts = []  # Zurücksetzen der Ebene 1 für jeden neuen Text
        self.levels[1].add_part(TextPart(text, TextTag.TEXT, 0))
        print(f"DEBUG: Inhalt von level 1 nach dem Hinzufügen: {[part.content for part in self.levels[1].parts]}")

    def identify_number_words(self) -> None:
        print("Identifiziere Zahlwörter")
        level_1 = self.levels[1]
        level_2 = self.levels[2]
        level_2.parts = []

        for part in level_1.parts:
            if part.tag == TextTag.TEXT:
                words = part.content.split()
                i = 0
                while i < len(words):
                    if any(indicator in words[i].lower() for indicator in NUMBER_INDICATORS):
                        start = i
                        while i < len(words) and (any(indicator in words[i].lower() for indicator in NUMBER_INDICATORS) or words[i].lower() == 'und'):
                            i += 1
                        # Text vor dem potenziellen Zahlwort
                        if start > 0:
                            level_2.add_part(TextPart(" ".join(words[:start]), TextTag.TEXT, self.current_position))
                            self.current_position += len(" ".join(words[:start])) + 1
                        # Potenzielles Zahlwort und Rest
                        level_2.add_part(TextPart(" ".join(words[start:]), TextTag.POSSIBLE_NUMBER, self.current_position))
                        self.current_position += len(" ".join(words[start:])) + 1
                        break
                    i += 1
                if i == len(words):  # Kein Zahlwort gefunden
                    level_2.add_part(TextPart(part.content, TextTag.TEXT, self.current_position))
                    self.current_position += len(part.content) + 1
            else:
                level_2.add_part(part)

        print(f"DEBUG: Ebene 2 nach Identifikation: {[(p.content, p.tag) for p in level_2.parts]}")

    def process_number_words(self) -> None:
        print("Verarbeite Zahlwörter")
        level_2 = self.levels[2]
        level_3 = self.levels[3]
        level_3.parts = []

        i = 0
        while i < len(level_2.parts):
            part = level_2.parts[i]
            if part.tag == TextTag.POSSIBLE_NUMBER:
                words = part.content.split()
                if len(words) == 1:
                    value, _ = parse_german_number(part.content)
                    if value is not None:
                        new_part = TextPart(str(value), TextTag.NUMBER, part.position)
                        new_part.value = value
                        level_3.add_part(new_part)
                    else:
                        level_3.add_part(TextPart(part.content, TextTag.TEXT, part.position))
                else:
                    processed_parts = self.process_word_pairs(words, part.position)
                    level_3.parts.extend(processed_parts)
                    # Wenn es einen Rest gibt, fügen wir ihn zurück zu level_2 hinzu
                    if processed_parts[-1].tag == TextTag.POSSIBLE_NUMBER:
                        level_2.parts.insert(i+1, processed_parts[-1])
                        level_3.parts.pop()  # Entferne den Rest aus level_3
            else:
                level_3.add_part(part)
            i += 1

        print(f"DEBUG: Ebene 3 nach Verarbeitung: {[(p.content, p.tag) for p in level_3.parts]}")

    def process_word_pairs(self, words: List[str], start_position: int) -> List[TextPart]:
        result = []
        i = 0
        while i < len(words):
            if i < len(words) - 1:
                pair = words[i:i+2]
                if pair[1].lower() in [word.lower() for word in LARGE_NUMBER_WORDS]:
                    value, _ = parse_german_number(" ".join(pair))
                    if value is not None:
                        part = TextPart(str(value), TextTag.LARGE_NUMBER, start_position + i)
                        part.value = value
                        result.append(part)
                        i += 2
                    else:
                        result.append(TextPart(pair[0], TextTag.TEXT, start_position + i))
                        i += 1
                elif pair[0].lower() in ['ein', 'eine'] and pair[1].lower() not in LARGE_NUMBER_WORDS:
                    result.append(TextPart(pair[0], TextTag.TEXT, start_position + i))
                    result.append(TextPart(pair[1], TextTag.TEXT, start_position + i + 1))
                    i += 2
                else:
                    value, original = parse_german_number(pair[0])
                    if value is not None:
                        result.append(TextPart(str(value), TextTag.NUMBER, start_position + i))
                        # Der Rest wird als mögliche Zahl zurück in die Verarbeitung gegeben
                        if i + 1 < len(words):
                            rest = " ".join(words[i+1:])
                            result.append(TextPart(rest, TextTag.POSSIBLE_NUMBER, start_position + i + 1))
                        return result
                    else:
                        result.append(TextPart(str(original), TextTag.TEXT, start_position + i))
                    i += 1
            else:
                value, original = parse_german_number(words[i])
                if value is not None:
                    result.append(TextPart(str(value), TextTag.NUMBER, start_position + i))
                else:
                    result.append(TextPart(str(original), TextTag.TEXT, start_position + i))
                i += 1
        return result

    def accumulate_numbers(self) -> None:
        print("Akkumuliere Zahlen")
        level_3 = self.levels[3]
        level_4 = self.levels[4]
        level_4.parts = []
        i = 0
        while i < len(level_3.parts):
            part = level_3.parts[i]
            if part.tag in [TextTag.LARGE_NUMBER, TextTag.NUMBER]:
                accumulated_parts, next_i = self.accumulate_number_sequence(level_3.parts, i)
                self.add_accumulated_to_level_4(accumulated_parts, level_4)

                # Füge alle Textteile zwischen dieser und der nächsten Zahl hinzu
                for j in range(i+1, next_i):
                    if level_3.parts[j].tag == TextTag.TEXT:
                        level_4.add_part(level_3.parts[j])

                i = next_i
            else:
                level_4.add_part(part)
                i += 1

    def accumulate_number_sequence(self, parts: List[TextPart], start_index: int) -> Tuple[List[TextPart], int]:
        print(f"Akkumuliere Zahlensequenz ab Index {start_index}")
        accumulated = [parts[start_index]]
        i = start_index + 1
        while i < len(parts):
            current_part = parts[i]
            if current_part.tag not in [TextTag.LARGE_NUMBER, TextTag.NUMBER]:
                break
            if self.should_accumulate(accumulated[-1], current_part):
                accumulated.append(current_part)
                i += 1
            else:
                break
        return accumulated, i

    def should_accumulate(self, prev_part: TextPart, current_part: TextPart) -> bool:
        if prev_part.value is None or current_part.value is None:
            return False
        if prev_part.tag == TextTag.LARGE_NUMBER and current_part.tag == TextTag.LARGE_NUMBER:
            return current_part.value < prev_part.value
        elif prev_part.tag == TextTag.LARGE_NUMBER and current_part.tag == TextTag.NUMBER:
            return True
        elif prev_part.tag == TextTag.NUMBER and current_part.tag == TextTag.LARGE_NUMBER:
            return False
        else:
            return prev_part.value > current_part.value

    def add_accumulated_to_level_4(self, accumulated_parts: List[TextPart], level_4: TextLevel):
        print(f"Füge akkumulierte Teile zu Ebene 4 hinzu: {[p.content for p in accumulated_parts]}")

        total_value = 0
        for part in accumulated_parts:
            if part.tag in [TextTag.LARGE_NUMBER, TextTag.NUMBER]:
                value, _ = parse_german_number(part.content)
                if value is not None:
                    total_value += value
                else:
                    print(f"Warnung: Konnte '{part.content}' nicht in eine Zahl umwandeln.")
            else:
                print(f"Warnung: Unerwartetes Tag {part.tag} für Teil '{part.content}'")

        new_part = TextPart(str(total_value), TextTag.TEXT, accumulated_parts[0].position)
        level_4.add_part(new_part)

    def reconstruct_text(self) -> str:
        sorted_parts = sorted(self.levels[4].parts, key=lambda x: x.position)
        result = ' '.join(part.content for part in sorted_parts)
        print(f"DEBUG: Rekonstruierter Text (vor Bereinigung): {result}")
        result = re.sub(r'\s+', ' ', result).strip()
        print(f"Rekonstruierter Text: {result}")
        return result

    def print_levels_state(self):
        for level, text_level in self.levels.items():
            print(f"DEBUG: Inhalt von Ebene {level}: {[(part.content, part.tag) for part in text_level.parts]}")

def parse_german_number(words):
    if isinstance(words, list):
        words = " ".join(words)
    words = str(words).strip()  # Ensure words is a string and remove leading/trailing whitespace
    print(f"Parse deutsche Zahl: {words}")

    # Prüfen, ob die Eingabe bereits eine Zahl ist
    try:
        return int(words), None
    except ValueError:
        pass  # Wenn es keine Zahl ist, fahren wir mit der Wortanalyse fort

    print(f"DEBUG: Startworte: {words}")

    # Schritt 1: Ziffern und Operatoren extrahieren
    tokens = []
    remaining_word = words.lower()
    while remaining_word:
        if remaining_word.startswith('und'):
            tokens.append('+')
            remaining_word = remaining_word[3:]
        else:
            found = False
            for key, val in sorted(GERMAN_NUMBER_DICT.items(), key=lambda x: len(x[0]), reverse=True):
                if remaining_word.startswith(key):
                    tokens.append(val)
                    remaining_word = remaining_word[len(key):]
                    found = True
                    break
            if not found:
                # Wenn kein bekanntes Zahlwort gefunden wurde, brechen wir ab
                return None, words

    print(f"DEBUG: Extrahierte Tokens: {tokens}")

    if not tokens or (len(tokens) == 1 and tokens[0] == '+'):
        return None, words

    # Schritt 2: Tokens gruppieren und Gleichung erstellen
    equation = []
    current_group = 0
    i = 0
    while i < len(tokens):
        token = tokens[i]
        if token == '+':
            if i + 2 < len(tokens) and tokens[i+2] >= 1000:
                # Behandlung von Fällen wie "zweiundvierzigtausend"
                current_group += tokens[i+1]
                equation.extend([str(current_group), '*', str(tokens[i+2]), '+'])
                current_group = 0
                i += 3
            else:
                if current_group:
                    equation.append(str(current_group))
                    current_group = 0
                equation.append('+')
                i += 1
        elif token >= 1000:
            if current_group:
                equation.extend([str(current_group), '*', str(token), '+'])
                current_group = 0
            else:
                equation.extend([str(token), '+'])
            i += 1
        elif token == 100:
            current_group *= token
            i += 1
        else:
            current_group += token
            i += 1

    if current_group:
        equation.append(str(current_group))

    # Entferne überschüssige Operatoren am Ende
    while equation and equation[-1] in ['+', '*']:
        equation.pop()

    equation_str = ' '.join(equation)
    print(f"DEBUG: Generierte Gleichung: {equation_str}")

    # Schritt 3: Gleichung auswerten
    if equation_str:
        result = eval(equation_str)
        print(f"DEBUG: Endergebnis: {result}")
        return result, None
    else:
        return None, words

def ziffern_zu_zahlwoerter(zahl):
    """
    Konvertiert eine Zahl in ihre deutsche Wortdarstellung.

    :param zahl: Die zu konvertierende Zahl (int)
    :return: Die Wortdarstellung der Zahl (str)
    """
    if not isinstance(zahl, int) or zahl < 0:
        raise ValueError("Die Eingabe muss eine nicht-negative ganze Zahl sein.")

    if zahl == 0:
        return "null"

    einer = ["", "ein", "zwei", "drei", "vier", "fünf", "sechs", "sieben", "acht", "neun"]
    zehner = ["", "zehn", "zwanzig", "dreißig", "vierzig", "fünfzig", "sechzig", "siebzig", "achtzig", "neunzig"]
    sonderformen = {11: "elf", 12: "zwölf", 16: "sechzehn", 17: "siebzehn"}
    grosse_zahlen = [(1000000000000, "Billion"), (1000000000, "Milliarde"), (1000000, "Million"), (1000, "tausend")]

    def bis_999(n):
        if n in sonderformen:
            return sonderformen[n]

        ergebnis = ""
        if n >= 100:
            ergebnis += einer[n // 100] + "hundert"
            n %= 100

        if n > 0:
            if n < 20 and n not in sonderformen:
                ergebnis += einer[n] + "zehn"
            else:
                if n % 10 != 0:
                    ergebnis += einer[n % 10]
                    if n > 20:
                        ergebnis += "und"
                ergebnis += zehner[n // 10]

        return ergebnis

    ergebnis = ""
    for wert, name in grosse_zahlen:
        if zahl >= wert:
            anzahl = zahl // wert
            ergebnis += bis_999(anzahl) + name
            if name != "tausend":
                ergebnis += "en" if anzahl > 1 else " "
            else:
                ergebnis += ""
            zahl %= wert
            if zahl > 0 and zahl < 100:
                ergebnis += "und"

    if zahl > 0:
        ergebnis += bis_999(zahl)

    # Spezielle Behandlung für "ein" am Anfang
    if ergebnis.startswith("ein") and len(ergebnis) > 3:
        ergebnis = "eine" + ergebnis[3:]

    return ergebnis.strip()

# Globale Variablen und Konstanten

GERMAN_NUMBER_DICT = {
    'null': 0, 'ein': 1, 'eine': 1, 'eins': 1, 'zwei': 2, 'drei': 3, 'vier': 4, 'fünf': 5,
    'sechs': 6, 'sieben': 7, 'acht': 8, 'neun': 9, 'zehn': 10, 'elf': 11, 'zwölf': 12,
    'dreizehn': 13, 'vierzehn': 14, 'fünfzehn': 15, 'sechzehn': 16, 'siebzehn': 17,
    'achtzehn': 18, 'neunzehn': 19, 'zwanzig': 20, 'dreißig': 30, 'vierzig': 40,
    'fünfzig': 50, 'sechzig': 60, 'siebzig': 70, 'achtzig': 80, 'neunzig': 90,
    'hundert': 100, 'tausend': 1000, 'million': 1000000, 'millionen': 1000000,
    'milliarde': 1000000000, 'milliarden': 1000000000
}

LARGE_NUMBER_WORDS = ["million", "millionen", "milliarde", "milliarden", "billion", "billionen", "billiarde", "billiarden", "trillion", "trillionen", "trilliarde", "trilliarden"]

NUMBER_INDICATORS = list(GERMAN_NUMBER_DICT.keys()) + LARGE_NUMBER_WORDS

# Hauptfunktionen

def words_to_digits(text: str) -> str:
    """
    Konvertiert Zahlwörter in einem Text zu Ziffern.

    :param text: Der zu konvertierende Text
    :return: Der Text mit konvertierten Ziffern
    """
    processor = TextProcessor()
    return processor.process_text(text)

def digits_to_words(text: str) -> str:
    """
    Konvertiert Ziffern in einem Text zu Zahlwörtern.

    :param text: Der zu konvertierende Text
    :return: Der Text mit konvertierten Zahlwörtern
    """
    def convert_number(match):
        number = int(match.group())
        return ziffern_zu_zahlwoerter(number)

    return re.sub(r'\b\d+\b', convert_number, text)

# Zusätzliche Erklärungen:

# 1. Die digits_to_words Funktion wurde aktualisiert, um die neue ziffern_zu_zahlwoerter Funktion zu verwenden.
#    Sie nutzt einen regulären Ausdruck, um alle Zahlen im Text zu finden und zu ersetzen.

# 2. Die ziffern_zu_zahlwoerter Funktion implementiert die komplexen Regeln der deutschen Zahlwortbildung.
#    Sie behandelt Zahlen bis zu einer Billion und berücksichtigt alle Sonderfälle und Unregelmäßigkeiten.

# 3. Die bestehende parse_german_number Funktion wurde beibehalten, da sie für die Umwandlung von Worten in Zahlen verwendet wird.

# 4. Die TextProcessor-Klasse und ihre Methoden wurden angepasst, um sicherzustellen, dass alle Teile des Textes
#    korrekt verarbeitet werden, einschließlich der Teile, die während der Verarbeitung zurück in niedrigere Ebenen verschoben werden.

# 5. Die process_text Methode wurde so angepasst, dass sie alle Verarbeitungsschritte nacheinander ausführt,
#    ohne zu versuchen, Teile zurück in Ebene 1 zu verschieben.

# 6. Es wäre sinnvoll, umfangreiche Tests für die neue Implementierung zu erstellen,
#    um sicherzustellen, dass sie alle möglichen Fälle korrekt behandelt.

# 7. In zukünftigen Versionen könnte man die Effizienz der Funktionen optimieren,
#    insbesondere für sehr große Zahlen oder häufige Aufrufe.
